requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
build==1.0.3
    # via pip-tools
click==8.1.7
    # via
    #   -r requirements.in
    #   pip-tools
packaging==23.2
    # via build
pillow==10.1.0
    # via -r requirements.in
pip-tools==7.3.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
wheel==0.41.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
